app/rvc_infer.py
import torch

class RVCModel:
    def __init__(self, model_path, index_path, device="cuda"):
        self.device = device
        self.model = self.load_model(model_path)
        self.index = self.load_index(index_path)

    def load_model(self, path):
        model = torch.load(path, map_location=self.device)
        model.eval()
        return model

    def load_index(self, path):
        import faiss
        return faiss.read_index(path)

    def convert(self, audio_np, pitch_shift=0, index_rate=0.75):
        # Placeholder — plug in real RVC pipeline
        # 1. Extract features (HuBERT)
        # 2. Apply index
        # 3. Run generator
        # 4. Return waveform
        return audio_np
