requirements.txt
fastapi
uvicorn
torch
numpy
soundfile
faiss-cpu
librosa
